+++ controller/ue/SharedState.py
@@ -1,0 +1,6 @@
+
+class SharedState:
+    ue_list = []
+    attack_args = []
+    cli_args = None
+    ue_index = 1
