+++ controller/ue/ue_controller.py
@@ -19,7 +19,6 @@ from kivy.uix.floatlayout import FloatLayout
 from kivy.uix.scrollview import ScrollView
 from kivy.clock import Clock
 from kivy.uix.image import Image
-from kivy.uix.rst import RstDocument
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
 import matplotlib.pyplot as plt
 import numpy as np
@@ -48,440 +47,8 @@ parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.insert(0, parent_dir)
 
 from ue_interface import Ue
-
-class LandingPage(Screen):
-
-    def __init__(self, **kwargs):
-        self.animation_completed = 0
-        super().__init__(**kwargs)
-
-        layout = RelativeLayout()
-        self.background = Image(source='Webimage.png', allow_stretch=True, keep_ratio=False)
-        layout.add_widget(self.background)
-        self.welcome_label = Label(
-            text="NTIA Soft T UE",
-            font_size='30sp',
-            halign='center',
-            valign='middle'
-        )
-    
-        self.welcome_label.bind(size=self.welcome_label.setter('text_size'))
-        layout.add_widget(self.welcome_label)
-        self.welcome_label.pos_hint = {'center_x': 0.5, 'center_y': 0.5}
-
-        self.add_widget(layout)
-
-    def on_enter(self):
-        Clock.schedule_once(self.animate_transition, 1)
-
-    def animate_transition(self, *args):
-        anim = Animation(opacity=0, duration=1)
-        anim.start(self.welcome_label)
-        anim.start(self.background) 
-
-        self.animation_completed = 1
-        anim.bind(on_complete=self.switch_to_processes)
-
-    def switch_to_processes(self, *args):
-        self.manager.current = 'processes'
-
-class ProcessesPage(Screen):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        global ue_index
-        self.ue_index = ue_index
-        layout = BoxLayout(orientation='vertical')
-        self.process_container = BoxLayout(
-            orientation="vertical", 
-            size_hint_y=None,
-        )
-        self.process_container.bind(minimum_height=self.process_container.setter('height'))
-        self.process_scroll = ScrollView()
-        self.process_scroll.add_widget(self.process_container)
-        layout.add_widget(self.process_scroll)
-
-        add_ue_button = Button(text='New UE', on_press=self.open_add_ue_popup, background_color=[0,1,0,1], size_hint_y=None)
-        layout.add_widget(add_ue_button)
-
-        self.config_file = ""
-        self.ue_type = "clean"
-
-        global ue_list
-        for ue in ue_list:
-            self.add_ue_log(ue["type"], ue["config"], [], ue["handle"])
-
-        self.add_widget(layout)
-
-    def open_add_ue_popup(self, instance):
-        content = BoxLayout(orientation='vertical')
-        self.config_label = Label(text="Selected file: None")
-        self.ue_type_label = Label(text=f"UE type: {self.ue_type}")
-
-        cancel_button = Button(text="Cancel", size_hint_y=None, height=50)
-        select_config_button = Button(text="Select config", size_hint_y=None, height=50)
-        add_button = Button(text="Add", size_hint_y=None, height=50)
-
-        ue_type_spinner = Spinner(
-            text='clean',
-            values=('clean', 'tester'),
-            size_hint=(None, None),
-            size=(200, 44)
-        )
-
-        button_wrapper = BoxLayout(orientation='horizontal')
-
-        content.add_widget(self.config_label)
-        content.add_widget(select_config_button)
-        content.add_widget(self.ue_type_label)
-        content.add_widget(ue_type_spinner)
-        button_wrapper.add_widget(cancel_button)
-        button_wrapper.add_widget(add_button)
-        content.add_widget(button_wrapper)
-
-        self.popup = Popup(title='Add New UE', content=content, size_hint=(0.5, 0.5))
-        cancel_button.bind(on_press=self.popup.dismiss)
-        add_button.bind(on_press=self.add_ue)
-        select_config_button.bind(on_press=self.open_select_file_popup)
-        ue_type_spinner.bind(text=self.set_ue_type)
-        self.popup.open()
-
-    def open_select_file_popup(self, instance):
-        content = BoxLayout(orientation='vertical')
-        filechooser = FileChooserListView(path='/home/')
-        button_layout = BoxLayout(size_hint_y=None, height=50)
-        select_button = Button(text="Select", size_hint_y=None, height=50)
-        cancel_button = Button(text="Cancel", size_hint_y=None, height=50)
-
-        button_layout.add_widget(select_button)
-        button_layout.add_widget(cancel_button)
-        content.add_widget(filechooser)
-        content.add_widget(button_layout)
-
-        popup = Popup(title='Choose a File', content=content, size_hint=(0.9, 0.9))
-
-        select_button.bind(on_press=lambda x: self.select_file(filechooser, popup))
-        cancel_button.bind(on_press=popup.dismiss)
-
-        popup.open()
-
-    def select_file(self, filechooser, popup):
-        selected = filechooser.selection
-        if selected:
-            self.config_file = selected[0]
-            self.config_label.text = f'Selected file: {selected[0]}'
-        popup.dismiss()
-
-    def set_ue_type(self, spinner, text):
-        self.ue_type_label.text = f'UE type: {text}'
-        self.ue_type = text
-
-    def add_ue(self, instance):
-        self.popup.dismiss()
-        new_ue = Ue(self.ue_index)
-        global attack_args
-
-        if self.ue_type == "clean":
-            new_ue.start([self.config_file])
-        else:
-            new_ue.start([self.config_file] + attack_args)
-            
-        global ue_list
-        ue_list.append({
-            'id': str(uuid.uuid4()),
-            'type': self.ue_type,
-            'config': self.config_file,
-            'handle': new_ue,
-            'index': self.ue_index
-        })
-
-        self.add_ue_log(self.ue_type, self.config_file, attack_args, new_ue)
-
-        self.config_file = ""
-        self.ue_type = "clean"
-        self.ue_index += 1
-
-    def collect_logs(self, label_ref, ue_ref, log_ref, title_ref):
-        label_ref.text = ue_ref.output
-        if not ue_ref.isRunning:
-            title_ref.color = [1,0,0,1]
-        elif not ue_ref.isConnected:
-            title_ref.color = [1,1,0,1]
-        else:
-            title_ref.color = [0,1,0,1]
-        log_ref.scroll_y = 0
-
-    def add_ue_log(self, ue_type, config, arguments, handle):
-        log_view = BoxLayout(
-            orientation="vertical",
-            size_hint_y=None,
-            height=500
-        )
-        new_ue_text = Label(
-            text=f"starting UE ({ue_type})...", 
-            size_hint_y=None,
-            font_size="15sp",
-            padding=[10,20,10,20]
-        )
-
-        content_label = Label(
-            text=f"sudo srsue {config}",
-            font_size="20sp",
-            padding=[10,20,10,20],
-        )
-
-        if self.ue_type == "tester":
-            content_label.text = f"sudo srsue {config} {arguments}"
-
-        Clock.schedule_interval(lambda dt: self.collect_logs(new_ue_text, handle, log_view, content_label), 1)
-        log_view.add_widget(content_label)
-        log_view.add_widget(new_ue_text)
-        self.process_container.add_widget(log_view)
-
-
-
-class AttacksPage(Screen):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.layout = BoxLayout(orientation='vertical')
-        attack_option = Spinner(
-            text='imsi_capture',
-            values=(
-                'sdu_fuzzing',
-                'cqi_manipulation',
-                'rrc_signal_flooding',
-                'rach_jamming',
-                'rach_replay',
-                'preamble_collision',
-                'rach_signal_flooding',
-                'imsi_capture',
-            ),
-            size_hint=(None, None),
-            size=(400, 30)
-        )
-        attack_option.bind(text=self.set_attack_type)
-        self.layout.add_widget(attack_option)
-        self.attack_type = "None"
-        self.add_widget(self.layout)
-        self.num_fuzzed_bits = 1
-        self.target_message = "All"
-        self.previous_doc = RstDocument(source=f"../../docs/attacks/imsi_capture.rst",)
-        self.layout.add_widget(self.previous_doc)
-        self.previous_spinners = []
-
-    def set_attack_type(self, spinner, text):
-        self.attack_type = text
-
-        for spinner in self.previous_spinners:
-            self.layout.remove_widget(spinner)
-        self.previous_spinners = []
-
-        if text == "sdu_fuzzing":
-            target_message = Spinner(
-                text='Target Message',
-                values=('All', 'rrcSetupRequest','rrcRegistrationRequest'),
-                size_hint=(None, None),
-                size=(400, 44)
-            )
-            target_message.bind(text=self.set_target_message)
-            bits_to_fuzz = Spinner(
-                text='Number of Bits to Fuzz',
-                values=[str(i) for i in range(1, 11)],
-                size_hint=(None, None),
-                size=(200, 44)
-            )
-            bits_to_fuzz.bind(text=self.set_fuzzed_bits)
-            self.layout.add_widget(target_message)
-            self.layout.add_widget(bits_to_fuzz)
-            self.previous_spinners.append(target_message)
-            self.previous_spinners.append(bits_to_fuzz)
-        if text == "CQI Manipulation":
-            cqi_value = Spinner(
-                text='CQI Value',
-                values=[str(i * 100) for i in range(10)],
-                size_hint=(None, None),
-                size=(200, 44)
-            )
-            cqi_value.bind(text=self.set_cqi_value)
-            self.layout.add_widget(cqi_value)
-            self.previous_spinners.append(cqi_value)
-
-        new_doc = RstDocument(source=f"../../docs/attacks/{text}.rst",)
-        self.layout.add_widget(new_doc)
-        if self.previous_doc:
-            self.layout.remove_widget(self.previous_doc)
-        self.previous_doc = new_doc
-
-
-    def set_target_message(self, spinner, text):
-        self.target_message = text
-        if text != "All":
-            global attack_args
-            attack_args = ["--rrc.sdu_fuzzed_bits", str(self.num_fuzzed_bits)
-                           , "--rrc.fuzz_target_message", self.target_message]
-            self.title.text = f"--rrc.sdu_fuzzed_bits {self.num_fuzzed_bits} --rrc.fuzz_target_message {self.target_message}"
-
-    def set_fuzzed_bits(self, spinner, text):
-        self.num_fuzzed_bits = int(text)
-        global attack_args
-        attack_args = ["--rrc.sdu_fuzzed_bits", str(self.num_fuzzed_bits)
-                       , "--rrc.fuzz_target_message", self.target_message]
-        self.title.text = f"--rrc.sdu_fuzzed_bits {self.num_fuzzed_bits} --rrc.fuzz_target_message {self.target_message}"
-
-    def set_cqi_value(self, spinner, text):
-        global attack_args
-        attack_args = ["--phy.cqi_max", text, "--phy.cqi_fixed", text]
-        self.title.text = f"--phy.cqi_max {text} --phy.cqi_fixed {text}"
-
-
-
-
-class ResultsPage(Screen):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.layout = GridLayout(cols=1, padding=[10,40,10,10], spacing=10,)
-        self.add_widget(self.layout)
-        self.rendered_ue_list = []
-
-    def init_results(self):
-        global ue_list
-        for ue_ref in ue_list:
-            if ue_ref['id'] in self.rendered_ue_list:
-                continue
-            self.rendered_ue_list.append(ue_ref["id"])
-            ue_results_block = BoxLayout()
-            ue_results_block.add_widget(self.create_graph("iperf", ue_ref))
-            ue_results_block.add_widget(self.create_graph("ping", ue_ref))
-            self.layout.add_widget(Label(text=f"UE{ue_ref['index']}", size_hint_y=None, height=20, font_size="30sp"))
-            self.layout.add_widget(ue_results_block)
-
-
-    def create_graph(self, graph_type, ue_ref):
-        plot = MeshLinePlot(color=[0,1,0,1])
-        plot.line_width = 6
-        graph = Graph(
-            xlabel='time (s)',
-            ylabel='Bandwidth (MBits/sec)',
-            x_ticks_minor=1,
-            x_ticks_major=5,
-            y_ticks_major=5,
-            y_ticks_minor=1,
-            y_grid_label=True,
-            x_grid_label=True,
-            padding=5,
-            xlog=False,
-            ylog=False,
-            x_grid=True,
-            y_grid=True,
-            ymin=0,
-            ymax=70,
-            xmin=0,
-            xmax=100
-        )
-        if graph_type == "ping":
-            graph.ylabel = "Latency (ms)"
-            graph.ymax = 150
-
-        graph.add_plot(plot)
-        xdata = list(range(100))
-        ydata = [0] * 100
-        if graph_type == "ping":
-            Clock.schedule_interval(lambda dt: self.update_points(plot, ue_ref["handle"].ping_client.output, xdata, ydata), 1)
-        else:
-            Clock.schedule_interval(lambda dt: self.update_points(plot, ue_ref["handle"].iperf_client.output, xdata, ydata), 1)
-
-        return graph
-
-
-    def update_points(self, plot, data_ref, xdata, ydata):
-        if len(data_ref) > 0:
-            new_y = data_ref[-1]
-            ydata.append(new_y)
-            if len(ydata) > len(xdata):
-                ydata.pop(0)
-
-        while len(xdata) > len(ydata):
-            xdata.pop(0)
-
-        plot.points = list(zip(xdata, ydata))
-
-class MainApp(App):
-
-    def build(self):
-        self.screen_manager = ScreenManager()
-
-        self.landing = LandingPage(name='landing')
-        self.processes = ProcessesPage(name='processes')
-        self.attacks = AttacksPage(name='attacks')
-        self.results = ResultsPage(name='results')
-
-        self.screen_manager.add_widget(self.landing)
-        self.screen_manager.add_widget(self.processes)
-        self.screen_manager.add_widget(self.attacks)
-        self.screen_manager.add_widget(self.results)
-
-        # Define the button colors
-        self.default_color = [1, 1, 1, 1]  # White
-        self.highlighted_color = [0, 1, 0, 1]  # Green
-
-        # Create a layout for the buttons on top
-        self.button_layout = BoxLayout(size_hint_y=None, height=50)
-
-        self.button_process_page = Button(text='Processes', on_press=self.switch_to_processes, background_color=self.highlighted_color)
-        self.button_attacks_page = Button(text='Attacks', on_press=self.switch_to_attacks, background_color=self.default_color)
-        self.button_results_page = Button(text='Results', on_press=self.switch_to_results, background_color=self.default_color)
-
-        self.button_layout.add_widget(self.button_process_page)
-        self.button_layout.add_widget(self.button_attacks_page)
-        self.button_layout.add_widget(self.button_results_page)
-
-        main_layout = BoxLayout(orientation='vertical')
-        main_layout.add_widget(self.button_layout)
-        main_layout.add_widget(self.screen_manager)
-
-        self.button_layout.opacity = 0 
-        animationStatus = LandingPage()
-
-        if animationStatus.animation_completed == 1:
-                self.button_layout.opacity = 1
-        else:
-            self.button_layout.opacity = 0  
-        Clock.schedule_once(self.load_navigation, 0.25)
-
-        return main_layout
-
-    def load_navigation(self, *args):
-        animate = Animation(opacity=1, duration=2)
-        animate.start(self.button_layout)
-
-    def switch_to_processes(self, instance):
-        self.screen_manager.current = 'processes'
-        self.update_button_colors(self.button_process_page)
-
-    def switch_to_attacks(self, instance):
-        self.screen_manager.current = 'attacks'
-        self.update_button_colors(self.button_attacks_page)
-
-    def switch_to_results(self, instance):
-        self.screen_manager.get_screen('results').init_results()
-        self.screen_manager.current = 'results'
-        self.update_button_colors(self.button_results_page)
-
-    def update_button_colors(self, active_button):
-        # Reset all buttons to the default color
-        self.button_process_page.background_color = self.default_color
-        self.button_results_page.background_color = self.default_color
-        self.button_attacks_page.background_color = self.default_color
-
-        # Highlight the active button
-        active_button.background_color = self.highlighted_color
-
-    def on_stop(self):
-        print("App is stopping...")
-        global cli_args
-        send_command(cli_args.ip, cli_args.port, "gnb:stop")
-        global ue_list
-        for ue in ue_list:
-            ue["handle"].stop()
+from MainApp import MainApp
+from SharedState import SharedState
 
 def parse():
     script_dir = pathlib.Path(__file__).resolve().parent
@@ -504,42 +71,34 @@ def parse():
 
 
 def main():
-    os.system("sudo kill -9 $(ps aux | awk '/iperf/ && !/awk/ {print $2}')") # kill iperf processes
-    os.system("sudo kill -9 $(ps aux | awk '/srsue/ && !/awk/ {print $2}')") # kill srsRAN UE processes
-    global cli_args
+    os.system("sudo echo 'sudo password'") # Get sudo access for later
     args = parse()
-    cli_args = args
+    SharedState.cli_args = args
     send_command(args.ip, args.port, f"gnb:start:{args.gnb_config}")
 
-    global ue_list
-    ue_list = list()
-    global attack_args
-    attack_args = list()
 
     options = None
     with open(str(args.config), 'r') as file:
         options = yaml.safe_load(file)
-    global ue_index
-    ue_index = 1
+    SharedState.ue_index = 1
 
-    time.sleep(0.5) # wait for namespace initialization
     for ue in options.get("ues", []):
         if not os.path.exists(ue["config_file"]):
             print(f"Error: File not found {ue[config_file]}")
             return 1
-        new_ue = Ue(ue_index)
+        new_ue = Ue(SharedState.ue_index)
         if ue['type'] == "tester":
             new_ue.start([ue["config_file"]] + ue["args"].split(" "))
         else:
             new_ue.start([ue["config_file"]])
-        ue_list.append({
+        SharedState.ue_list.append({
             'id': str(uuid.uuid4()),
             'type': ue['type'],
             'config': ue['config_file'],
             'handle': new_ue,
-            'index': ue_index
+            'index': SharedState.ue_index
         })
-        ue_index += 1
+        SharedState.ue_index += 1
 
 
     MainApp().run()
